# apps/go-api/openapi.yaml
openapi: 3.0.3
info:
  title: Next-Go API
  version: 1.0.0
  description: Go API for Next.js monorepo template
servers:
  - url: http://localhost:8080
    description: Local development
paths:
...
  /auth/oauth/callback:
    post:
      summary: OAuth callback from provider
      operationId: oauthCallback
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OAuthCallbackRequest"
      responses:
        "200":
          description: Existing user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthCallbackResponse"
        "201":
          description: New user created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthCallbackResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      required: [id, email, created_at]
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
          format: email
        name:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
...
    OAuthCallbackRequest:
      type: object
      required: [provider, provider_account_id, email]
      properties:
        provider:
          type: string
        provider_account_id:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
    OAuthCallbackResponse:
      type: object
      required: [user, is_new_user]
      properties:
        user:
          $ref: "#/components/schemas/User"
        is_new_user:
          type: boolean
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...